Refactor: moved specific functions to own files
